pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tidywizard"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "streamlit>=1.36",
  "pandas>=2.2",
  "numpy>=1.26",
  "pyyaml>=6.0",
  "openpyxl>=3.1",
  "pytest>=7",
  "ruff>=0.1",
  "mypy>=1.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.setuptools.packages.find]
where = ["."]
include = ["core*"]
